Added script for histogram

+++ histogram/catapult/script/impl_gen.py
@@ -20,53 +20,45 @@ try:
 except:
     pass
 def run_script(path):
-    print(path, "started at", datetime.datetime.now())
-    try:
-        m = re.search('histogram(\d+)', path)
-        if os.path.isdir('impl/'+m[0]):
-            pass
-        else:
-            os.makedirs('impl/'+m[0])
-        shutil.copy(path+'/'+'rtl.v','impl/'+m[0]+'/')
-        shutil.copy(path+'/'+'concat_rtl.v','impl/'+m[0]+'/')
-        shutil.copy('vivado_mentor.tcl','impl/'+m[0]+'/')
-        shutil.copy('extraction.tcl','impl/'+m[0]+'/')
-        shutil.copy('constr.xdc','impl/'+m[0]+'/')
-        '''
-        with open('impl/'+m[0]+'/'+'vivado_mentor.tcl', "r") as fin:
-            text=''
-            for line in fin:
-                newline=line
-                newline=line.replace("%1",'impl/'+m[0])
-                text=text+newline
-        with open('impl/'+m[0]+'/'+'vivado_mentor.tcl', "w") as fin:
-                fin.writelines(text)
-        '''
-        start = time.time()
-        command = " ".join(["timeout 7200","vivado","-mode","batch","-source","./vivado_mentor.tcl"])
-        subprocess.check_output(command, cwd='impl/'+m[0], shell=True)
-        end = time.time()
+    if os.path.isfile(path+'/'+'rtl.v'):
+        print(path, "started at", datetime.datetime.now())
+        try:
+            m = re.search('histogram(\d+)', path)
+            if os.path.isdir('impl/'+m[0]):
+                pass
+            else:
+                os.makedirs('impl/'+m[0])
+            shutil.copy(path+'/'+'rtl.v','impl/'+m[0]+'/')
+            shutil.copy(path+'/'+'concat_rtl.v','impl/'+m[0]+'/')
+            shutil.copy('vivado_mentor.tcl','impl/'+m[0]+'/')
+            shutil.copy('extraction.tcl','impl/'+m[0]+'/')
+            shutil.copy('constr.xdc','impl/'+m[0]+'/')
+            start = time.time()
+            command = " ".join(["timeout 7200","vivado","-mode","batch","-source","./vivado_mentor.tcl"])
+            subprocess.check_output(command, cwd='impl/'+m[0], shell=True)
+            end = time.time()
 
-        outFile = open(logName, 'at')
-        outFile.write(path + "\n")
-        outFile.close()
+            outFile = open(logName, 'at')
+            outFile.write(path + "\n")
+            outFile.close()
 
-        outFile = open(logName + ".time", 'at')
-        outFile.write(path + ',' + str(end-start) + '\n')
-        outFile.close()
+            outFile = open(logName + ".time", 'at')
+            outFile.write(path + ',' + str(end-start) + '\n')
+            outFile.close()
 
-        print(path, "end at", datetime.datetime.now(), " elapsed", end-start)
+            print(path, "end at", datetime.datetime.now(), " elapsed", end-start)
 	
-    except subprocess.CalledProcessError:
-        print(path, "Timeout at", datetime.datetime.now())
+        except subprocess.CalledProcessError:
+            print(path, "Timeout at", datetime.datetime.now())
 
-        outFile = open(logName + '.timeout', 'at')
-        outFile.write(path + '\n')
-        outFile.close()
+            outFile = open(logName + '.timeout', 'at')
+            outFile.write(path + '\n')
+            outFile.close()
 
-    except:
-        raise
-
+        except:
+            raise
+    else:
+        pass
 
 def main():
     if os.path.isfile(logName):
